[sailfish-crypto] remove included cpp file, contributes to JB#40058
