+++ plugins/sqlcipherplugin/sqlcipherplugin.h
@@ -16,6 +16,7 @@
 #include "Crypto/key.h"
 #include "Crypto/result.h"
 
+#include "opensslcryptoplugin.h"
 #include "database_p.h"
 
 #include <QObject>
@@ -215,8 +216,6 @@ public:
 private:
     static QString databaseDirPath(bool isTestPlugin, const QString &databaseSubdir);
     void init_aes_encryption();
-    QByteArray aes_encrypt_plaintext(Sailfish::Crypto::CryptoManager::BlockMode blockMode, const QByteArray &plaintext, const QByteArray &key, const QByteArray &init_vector);
-    QByteArray aes_decrypt_ciphertext(Sailfish::Crypto::CryptoManager::BlockMode blockMode, const QByteArray &ciphertext, const QByteArray &key, const QByteArray &init_vector);
     Sailfish::Secrets::Result openCollectionDatabase(const QString &collectionName, const QByteArray &key, bool createIfNotExists);
     QMap<QString, Sailfish::Secrets::Daemon::Sqlite::Database *> m_collectionDatabases;
 
@@ -234,6 +233,8 @@ private:
         quint64 clientId = 0;
     };
     QMap<QTimer *, CipherSessionLookup> m_cipherSessionTimeouts;
+    Sailfish::Crypto::Daemon::Plugins::OpenSslCryptoPlugin m_opensslCryptoPlugin;
+    friend class Sailfish::Crypto::Daemon::Plugins::OpenSslCryptoPlugin;
 };
 
 } // namespace Plugins
