+++ plugins/sqlcipherplugin/sqlcipherplugin.pro
@@ -8,21 +8,21 @@ include($$PWD/../../lib/libsailfishsecrets.pri)
 include($$PWD/../../lib/libsailfishcrypto.pri)
 include($$PWD/../../database/database.pri)
 
-INCLUDEPATH += $$PWD/../opensslcryptoplugin/evp/
-DEPENDPATH += $$PWD/../opensslcryptoplugin/evp/
+INCLUDEPATH += . $$PWD/../opensslcryptoplugin $$PWD/../opensslcryptoplugin/evp/
+DEPENDPATH += . $$PWD/../opensslcryptoplugin $$PWD/../opensslcryptoplugin/evp/
 
 HEADERS += \
     $$PWD/../opensslcryptoplugin/evp/evp_p.h \
+    $$PWD/../opensslcryptoplugin/evp/evp_helpers_p.h \
+    $$PWD/../opensslcryptoplugin/opensslcryptoplugin.h \
     $$PWD/sqlcipherplugin.h
 
 SOURCES += \
     $$PWD/../opensslcryptoplugin/evp/evp.c \
+    $$PWD/../opensslcryptoplugin/opensslcryptoplugin.cpp \
     $$PWD/sqlcipherplugin.cpp \
     $$PWD/encryptedstorageplugin.cpp \
     $$PWD/cryptoplugin.cpp
-
-OTHER_FILES += \
-    $$PWD/../opensslcryptoplugin/cryptoplugin_common.cpp
 
 target.path=/usr/lib/Sailfish/Secrets/
 INSTALLS += target
